Move common message model interface to concern
